+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add support for multi-result responses of any formats
+
 ## [0.1.0] - 2025-05-26
 
 - Initial release